+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The changelog is available [on GitHub][2].
   Drop support of GHC-8.2.2.
 * [#252](https://github.com/kowainik/tomland/issues/252):
   Move to `hspec-*` family of libraries from `tasty-*`.
+* [#261](https://github.com/kowainik/tomland/issues/261):
+  Implement `tableMap` codec to use TOML keys as `Map` keys.
 * [#256](https://github.com/kowainik/tomland/issues/256):
   Rename modules to simplify module structure.
   __TODO__: write migration guide